separate home and admin components to different files
